Rewrite repository/test_repository_workflow as integrationtest.
